Fix critical bugs in speed calculation, trajectory visualization, and JSON serialization
Major bug fixes for production-ready bowling shot tracking:

1. Speed Calculation Fix (Priority: Critical)
   - Fixed incorrect speed calculation that was summing pixel-to-pixel distances
   - Changed to use known lane length (60 feet) / time elapsed
   - Optimized for head-mounted camera where pixel distance is unreliable
   - Speed results now show realistic bowling speeds (15-20 mph range)
   - Updated: src/speed_calculator.py lines 243-298

2. Trajectory Visualization Fix (Priority: Critical)
   - Fixed trajectory chart showing unreadable tiny dots at bottom
   - Implemented proper video-to-lane coordinate mapping
   - Trajectories now centered around board 20 with realistic spread
   - Hook points properly mapped using same transformation
   - Visual representation now matches actual ball path on lane
   - Updated: src/dashboard.py lines 127-222

3. JSON Serialization Fix (Priority: High)
   - Fixed numpy int64/float64 serialization errors in trajectory data
   - Added recursive convert_numpy_types() function
   - Handles all numpy types in trajectory points and analysis data
   - Prevents database save failures
   - Updated: src/video_processor.py lines 280-310

4. Python 3.13 Compatibility (Priority: High)
   - Updated numpy requirement from pinned 1.24.3 to >=1.26.0
   - Fixed Kalman filter dtype issues with explicit np.float32 casting
   - Removed Unicode checkmark characters causing Windows terminal errors
   - All print statements now use ASCII-safe [OK] markers
   - Updated: requirements.txt, src/ball_tracker.py, src/video_processor.py

5. Documentation Updates
   - Updated README.md with Python 3.13 support
   - Corrected performance expectations based on real-world testing
   - Updated speed calculation description to reflect new method
   - Added realistic processing time estimates (30 min per batch)

Technical improvements:
- Lane-length-based speed calculation eliminates calibration errors
- Proper coordinate normalization for trajectory visualization
- Comprehensive numpy type handling for database persistence
- Windows terminal compatibility with ASCII output
- Updated performance metrics to reflect actual processing times

+++ src/ball_tracker.py
@@ -373,16 +373,16 @@ class BallTracker:
 
         # Update Kalman filter
         if self.enable_kalman and self.kalman is not None:
-            measurement = np.array([[np.float32(best.x)], [np.float32(best.y)]])
+            measurement = np.array([[np.float32(best.x)], [np.float32(best.y)]], dtype=np.float32)
 
             if self.last_detection is None:
                 # First detection - initialize state
                 self.kalman.statePre = np.array([
                     [np.float32(best.x)],
                     [np.float32(best.y)],
-                    [0],
-                    [0]
-                ])
+                    [np.float32(0)],
+                    [np.float32(0)]
+                ], dtype=np.float32)
 
             self.kalman.correct(measurement)
             self.last_detection = (best.x, best.y)
